Add support for changing the internal tag name
Previously, the `InternalTag` scheme hard-coded the tag name to be "Case". This
commit makes the tag name configurable, but codes "Case" as the default, which
should make this change fully backwards-compatible. This is done by adding a new
argument to the set of `JsonFSharpConverter` constructors, along with a type
alias to clarify what the argument does.

The original test suite
passes, and four more test cases are added to be sure the new functionality
works.